Struct page_table_entry::aarch64::DescriptorAttr

pub struct DescriptorAttr(/* private fields */);

Available on AArch64 only.

Memory attribute fields in the VMSAv8-64 translation table format descriptors.

Implementations

impl DescriptorAttr

pub const VALID: Self = _

Whether the descriptor is valid.

pub const NON_BLOCK: Self = _

The descriptor gives the address of the next level of translation table or 4KB page. (not a 2M, 1G block)

pub const ATTR_INDX: Self = _

Memory attributes index field.

pub const NS: Self = _

Non-secure bit. For memory accesses from Secure state, specifies whether the output address is in Secure or Non-secure memory.

pub const AP_EL0: Self = _

Access permission: accessable at EL0.

pub const AP_RO: Self = _

Access permission: read-only.

pub const INNER: Self = _

Shareability: Inner Shareable (otherwise Outer Shareable).

pub const SHAREABLE: Self = _

Shareability: Inner or Outer Shareable (otherwise Non-shareable).

pub const AF: Self = _

The Access flag.

pub const NG: Self = _

The not global bit.

pub const CONTIGUOUS: Self = _

Indicates that 16 adjacent translation table entries point to contiguous memory regions.

pub const PXN: Self = _

The Privileged execute-never field.

pub const UXN: Self = _

The Execute-never or Unprivileged execute-never field.

pub const PXN_TABLE: Self = _

PXN limit for subsequent levels of lookup.

pub const XN_TABLE: Self = _

XN limit for subsequent levels of lookup.

pub const AP_NO_EL0_TABLE: Self = _

Access permissions limit for subsequent levels of lookup: access at EL0 not permitted.

pub const AP_NO_WRITE_TABLE: Self = _

Access permissions limit for subsequent levels of lookup: write access not permitted.

pub const NS_TABLE: Self = _

For memory accesses from Secure state, specifies the Security state for subsequent levels of lookup.

impl DescriptorAttr

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl DescriptorAttr

pub const fn iter(&self) -> Iter<DescriptorAttr>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<DescriptorAttr>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl DescriptorAttr

pub const fn from_mem_attr(idx: MemAttr) -> Self

Constructs a descriptor from the memory index, leaving the other fields empty.

pub const fn mem_attr(&self) -> Option<MemAttr>

Returns the memory attribute index field.