Crate page_table_entry

This crate provides the definition of page table entry for various hardware architectures.

Currently supported architectures and page table entry types:

x86: x86_64::X64PTE

ARM: aarch64::A64PTE

RISC-V: riscv::Rv64PTE

All these types implement the GenericPTE trait, which provides unified methods for manipulating various page table entries.

Modules

aarch64AArch64

AArch64 VMSAv8-64 translation table format descriptors.

riscvRISC-V RV32 or RISC-V RV64

RISC-V page table entries.

x86_64x86-64

x86 page table entries on 64-bit paging.

Structs

MappingFlags

Generic page table entry flags that indicate the corresponding mapped memory region permissions and attributes.

Traits

GenericPTE

A generic page table entry.